refactored the login and signup pages
