+++ package.json
@@ -11,7 +11,6 @@
     "axios": "^0.21.1",
     "bootstrap": "^5.1.0",
     "chart.js": "^3.5.1",
-    "chartist": "^0.11.4",
     "mdbreact": "^5.0.2",
     "node-sass": "4.14.1",
     "primeflex": "^2.0.0",
