+++ src/component/login/Login.js
@@ -1,17 +1,25 @@
 import React, { useState } from "react";
 import Form from "react-bootstrap/Form";
 import Button from "react-bootstrap/Button";
-import "./Login.css";
+import './Login.css'
 import axios from "axios";
-import clapp from '../../resources/1_clap_footer.png';
-import { MDBCol, MDBContainer, MDBRow} from "mdbreact";
 import photo from "../../resources/phone.png";
+import { Row, Col } from 'react-bootstrap';
+import { MDBCol, MDBContainer, MDBRow} from "mdbreact";
 import PlaystoreImg from '../../resources/playstore.png';
 import AppleImg from '../../resources/apple-logo.png';
+import clapp from '../../resources/1_clap_footer.png';
 
-export default function Login() {
+export default function Signup() {
+  const [fname, setFName] = useState("");
+  const [lname, setLName] = useState("");
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
+  const [address, setAddress] = useState("");
+  const [cpassword, setcPassword] = useState("");
+  const [pincode, setPincode] = useState("");
+  const [phoneNo, setPhoneNo] = useState("");
+  
 
   function validateForm() {
     return email.length > 0 && password.length > 0;
@@ -28,46 +36,122 @@ export default function Login() {
   }
 
   return (
-    <div>
+
+
     <MDBContainer>
+<br/>
+    <br/>
+     
     <MDBRow >
-    <MDBCol md="2" style={{textAlign:"center"}}>
+    <div style={{textAlign:"center"}}>
+    <h1 style={{display:"inline-block"}} >Sign Up for </h1>
+    <img src={clapp} alt="1Clapp_image" width="140" height="50" style={{marginLeft:"10px",marginTop:"-24px"}}  />
+
+    </div>
+    <MDBCol md="5" style={{textAlign:"center"}}>
+    <br/>
       <br/>
       <br/>
-      <br/>
-      <br/>
-      <br/>
-      <br/>
-    <h5>Download the 1Clapp App</h5>
+ 
+    
+    <h2>One Stop Solution For Everyone's Needs</h2>
+    <br/>
+    <div style={{textAlign:"center"}}><img src={photo} alt="phone_image" className="image-phone-login"/></div>
+    <br/>
     <button className="btn btn-sm btn-light butn" style={{borderRadius:"10px"}}>  <img src={PlaystoreImg} alt="" className="button_img " />GOOGLE PLAY </button> 
     <button className="btn btn-sm btn-light butn" style={{borderRadius:"10px"}}>  <img src={AppleImg} alt="" className="button_img " />APPLE STORE </button>
     </MDBCol>
-      <MDBCol md="3" >
-      <br/>
-      <br/>
-      <br/>
-      <br/>
-      <div style={{textAlign:"right"}}><img src={photo} alt="phone_image" className="image-phone-login"/></div>
-       </MDBCol>
-          <MDBCol md="7">
-    <div className="Login">
-     <h1>Login</h1>
-
-      <Form onSubmit={handleSubmit} style={{borderRadius:"3%"}}>
-      <div style={{textAlign:"center"}}>
-      <img src={clapp} alt="1Clapp_image" width="140" height="50"/>
-      </div>  
-    
-      <br/>
-        <Form.Group size="lg" controlId="email" >
-          <Form.Label>Email</Form.Label>
+   
+   
+    <MDBCol md="7" >
+    <div className="SignUp">
+      <Form onSubmit={handleSubmit}>
+      <Row className="g-2">
+      <Col md>
+      <Form.Group size="lg" controlId="fname">
+          <Form.Label >First Name</Form.Label>
           <Form.Control
             autoFocus
+            type="name"
+            value={fname}
+            onChange={(e) => setFName(e.target.value)}
+          />
+          
+          </Form.Group>
+          </Col>
+          <Col md>
+          <Form.Group size="lg" controlId="lname">
+          <Form.Label >Last Name</Form.Label>
+          <Form.Control
+            autoFocus
+            type="name"
+            value={lname}
+            onChange={(e) => setLName(e.target.value)}
+          />
+          </Form.Group>
+          </Col>
+          </Row>
+          <Row className="g-1">
+          <Col md>
+          <Form.Group size="lg" controlId="address">
+          <Form.Label >Address</Form.Label>
+          <Form.Control
+          as="textarea"
+            type="textarea"
+            value={address}
+            onChange={(e) => setAddress(e.target.value)}
+          />
+          </Form.Group>
+          </Col>
+          </Row>
+          <Row className="g-2">
+          <Col md>
+          <Form.Group size="lg" controlId="pincode">
+          <Form.Label >Pincode</Form.Label>
+          <Form.Control
+            type="text"
+            value={pincode}
+            onChange={(e) => setPincode(e.target.value)}
+          />
+          </Form.Group>
+          </Col>
+          <Col md>
+          <Form.Group size="lg" controlId="phoneNo">
+          <Form.Label >Phone Number</Form.Label>
+          <Form.Control
+            type="text"
+            value={phoneNo}
+            onChange={(e) => setPhoneNo(e.target.value)}
+          />
+          </Form.Group>
+          </Col>
+          </Row>
+        <Form.Group size="lg" controlId="email">
+          <Form.Label>Email</Form.Label>
+          <Form.Control
+        
             type="email"
             value={email}
             onChange={(e) => setEmail(e.target.value)}
           />
         </Form.Group>
+        <Form.Group size="lg" controlId="upi">
+          <Form.Label>UPI/PAYTM</Form.Label>
+          <select  style={{marginLeft:"10px"}}>
+          <option selected value="yes">YES</option>
+          <option value="no">NO</option>
+          </select>
+         
+          </Form.Group>
+          <Form.Group size="lg" controlId="service">
+          <Form.Label >CHOOSE SERVICE</Form.Label>
+          <select style={{marginLeft:"10px"}}>
+          <option selected value="platinum">PLATINUM</option>
+          <option value="gold">GOLD</option>
+          <option value="silver">SILVER</option>
+          </select>
+          <a href="/servicesSGP" style={{marginLeft:"10px"}}>?</a>
+          </Form.Group>
         <Form.Group size="lg" controlId="password">
           <Form.Label>Password</Form.Label>
           <Form.Control
@@ -76,19 +160,23 @@ export default function Login() {
             onChange={(e) => setPassword(e.target.value)}
           />
         </Form.Group>
-        <Button block size="md " type="submit" disabled={!validateForm()} className="btn btn-warning">
-          Login
+        <Form.Group size="lg" controlId="password">
+          <Form.Label>Confirm Password</Form.Label>
+          <Form.Control
+            type="password"
+            value={cpassword}
+            onChange={(e) => setcPassword(e.target.value)}
+          />
+        </Form.Group>
+        <Button block size="lg" type="submit" className="btn-warning" disabled={!validateForm()}>
+          Register
         </Button>
-        <a href="/signup">Forgot Password ?</a>
-        <br/>
-        <hr/>
-        <div><h6>Dont have an account? <a href="/signup">Sign Up</a></h6> </div>
       </Form>
+      </div>
+      </MDBCol>
+      </MDBRow>
+      </MDBContainer>
  
-    </div>
-    </MDBCol>
-    </MDBRow>
-    </MDBContainer>
-    </div>
   );
 }
+
